requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
pydantic[email]==2.7.1
email-validator==2.1.1
python-dotenv==1.0.1
passlib[argon2]==1.7.4

tensorflow-cpu==2.16.1
numpy==1.26.4

streamlit==1.35.0
requests==2.31.0
pandas==2.2.2